+++ MatrixKit/Models/MXKAppSettings.h
@@ -101,6 +101,11 @@
  */
 @property (nonatomic) BOOL hidePreJoinedUndecryptableEvents;
 
+/**
+ Indicate to hide un-decryptable events in the room. Default is `YES`.
+ */
+@property (nonatomic) BOOL hideUndecryptableEvents;
+
 #pragma mark - Room members
 
 /**